requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=9.0.0
moderngl>=5.6.0
pygame>=2.0.0
PyOpenGL>=3.1.5
PyGLM>=2.5.0
scipy>=1.7.0
matplotlib>=3.5.0
